# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Swordfish Email Security API
  description: |
    API for Swordfish - Enterprise email security platform with ML-powered threat detection.

    ## Authentication
    All endpoints require authentication via Clerk. Include the session token in the Authorization header.

    ## Multi-tenancy
    Requests are automatically scoped to the authenticated user's organization (tenant).

    ## Rate Limits
    - Standard tier: 100 requests/minute
    - Enterprise tier: 1000 requests/minute
  version: 1.0.0
  contact:
    name: Swordfish Support
    email: support@swordfish.security
  license:
    name: Proprietary
    url: https://swordfish.security/license
servers:
  - url: https://api.swordfish.security/v1
    description: Production server
  - url: https://staging-api.swordfish.security/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development
...
paths:
  /threats:
    get:
      tags: [Threats]
      summary: List detected threats
      description: Returns a paginated list of detected threats (quarantined/blocked emails)
      operationId: listThreats
      parameters:
        - name: status
          in: query
          description: Filter by threat status
          schema:
            type: string
            enum: [quarantine, block, suspicious, pass]
        - name: severity
          in: query
          description: Minimum severity level
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: startDate
          in: query
          description: Filter from date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter to date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of threats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /threats/{id}:
    get:
      tags: [Threats]
      summary: Get threat details
      description: Returns detailed information about a specific threat
      operationId: getThreat
      parameters:
        - $ref: '#/components/parameters/ThreatId'
      responses:
        '200':
          description: Threat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threat'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Threats]
      summary: Delete a threat
      description: Permanently deletes a threat record
      operationId: deleteThreat
      parameters:
        - $ref: '#/components/parameters/ThreatId'
      responses:
        '200':
          description: Threat deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /threats/bulk:
    post:
      tags: [Threats]
      summary: Bulk threat operations
      description: Perform actions on multiple threats at once
      operationId: bulkThreatAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, action]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                action:
                  type: string
                  enum: [release, delete, block_sender]
      responses:
        '200':
          description: Bulk operation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed:
                    type: integer
                  failed:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
  /analytics:
    get:
      tags: [Analytics]
      summary: Get dashboard statistics
      description: Returns aggregated security statistics for the dashboard
      operationId: getDashboardStats
      parameters:
        - name: period
          in: query
          description: Time period for statistics
          schema:
            type: string
            enum: [24h, 7d, 30d, 90d]
            default: 7d
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
  /analytics/timeseries:
    get:
      tags: [Analytics]
      summary: Get time series data
      description: Returns threat detection data over time
      operationId: getTimeSeries
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [24h, 7d, 30d, 90d]
        - name: metric
          in: query
          description: Metric to retrieve
          schema:
            type: string
            enum: [threats, scanned, blocked, quarantined]
            default: threats
      responses:
        '200':
          description: Time series data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSeriesResponse'
...
  /policies/allowlist:
    get:
      tags: [Policies]
      summary: List allowlist entries
      operationId: listAllowlist
      responses:
        '200':
          description: Allowlist entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AllowlistEntry'

    post:
      tags: [Policies]
      summary: Add to allowlist
      operationId: addToAllowlist
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAllowlistEntry'
      responses:
        '200':
          description: Entry added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowlistEntry'
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token
    webhookToken:
      type: apiKey
      in: header
      name: X-Webhook-Token
      description: Webhook authentication token
  parameters:
    ThreatId:
      name: id
      in: path
      required: true
      description: Unique threat identifier
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    Threat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
        from:
          type: string
          format: email
        to:
          type: array
          items:
            type: string
            format: email
        subject:
          type: string
        verdict:
          type: string
          enum: [pass, suspicious, quarantine, block]
        score:
          type: integer
          minimum: 0
          maximum: 100
        confidence:
          type: number
          minimum: 0
          maximum: 1
        signals:
          type: array
          items:
            $ref: '#/components/schemas/Signal'
        explanation:
          type: string
        recommendation:
          type: string
        createdAt:
          type: string
          format: date-time
    Signal:
      type: object
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        score:
          type: number
        category:
          type: string
        detail:
          type: string
    ThreatListResponse:
      type: object
      properties:
        threats:
          type: array
          items:
            $ref: '#/components/schemas/Threat'
        stats:
          type: object
          properties:
            total:
              type: integer
            quarantined:
              type: integer
            blocked:
              type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    DashboardStats:
      type: object
      properties:
        stats:
          type: object
          properties:
            totalScanned:
              type: integer
            threatsBlocked:
              type: integer
            threatsQuarantined:
              type: integer
            passedEmails:
              type: integer
            avgProcessingTime:
              type: number
            detectionRate:
              type: number
    TimeSeriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              value:
                type: number
        period:
          type: string
        metric:
          type: string
...
    AllowlistEntry:
      type: object
      properties:
        id:
          type: string
        pattern:
          type: string
        patternType:
          type: string
          enum: [email, domain, ip]
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
    CreateAllowlistEntry:
      type: object
      required: [pattern, patternType]
      properties:
        pattern:
          type: string
        patternType:
          type: string
          enum: [email, domain, ip]
        name:
          type: string
...
security:
  - clerkAuth: []